Refactor Invoker
Invoker implementation currently contains logic
to handle both config methods and also test methods.
This makes the Invoker class complex to understand.
Split the Invoker into ConfigInvoker and TestInvoker
and had the Invoker just use these two package
level classes for functionality.
