+++ src/main/java/org/testng/internal/TestMethodWithDataProviderMethodWorker.java
@@ -21,19 +21,18 @@ public class TestMethodWithDataProviderMethodWorker implements Callable<List<ITe
   private final ITestNGMethod[] m_beforeMethods;
   private final ITestNGMethod[] m_afterMethods;
   private final ConfigurationGroupMethods m_groupMethods;
-  private final Invoker m_invoker;
   private final ITestContext m_testContext;
   private int m_parameterIndex;
   private boolean m_skipFailedInvocationCounts;
   private int m_invocationCount;
   private final ITestResultNotifier m_notifier;
+  private final TestInvoker m_testInvoker;
 
   private final List<ITestResult> m_testResults = Lists.newArrayList();
   private int m_failureCount;
 
   public TestMethodWithDataProviderMethodWorker(
-      Invoker invoker,
-      ITestNGMethod testMethod,
+      TestInvoker testInvoker, ITestNGMethod testMethod,
       int parameterIndex,
       Object[] parameterValues,
       Object instance,
@@ -47,7 +46,7 @@ public class TestMethodWithDataProviderMethodWorker implements Callable<List<ITe
       int invocationCount,
       int failureCount,
       ITestResultNotifier notifier) {
-    m_invoker = invoker;
+    this.m_testInvoker = testInvoker;
     m_testMethod = testMethod;
     m_parameterIndex = parameterIndex;
     m_parameterValues = parameterValues;
@@ -70,11 +69,11 @@ public class TestMethodWithDataProviderMethodWorker implements Callable<List<ITe
     long start = System.currentTimeMillis();
     XmlSuite suite = m_testContext.getSuite().getXmlSuite();
 
-    final Invoker.FailureContext failure = new Invoker.FailureContext();
+    final ITestInvoker.FailureContext failure = new ITestInvoker.FailureContext();
     failure.count = m_failureCount;
     try {
       tmpResults.add(
-          m_invoker.invokeTestMethod(
+          m_testInvoker.invokeTestMethod(
               m_instance,
               m_testMethod,
               m_parameterValues,
@@ -95,7 +94,7 @@ public class TestMethodWithDataProviderMethodWorker implements Callable<List<ITe
           List<ITestResult> retryResults = Lists.newArrayList();
 
           m_failureCount =
-              m_invoker.retryFailed(
+              m_testInvoker.retryFailed(
                       instance,
                       m_testMethod,
                       m_parameterValues,
@@ -130,7 +129,7 @@ public class TestMethodWithDataProviderMethodWorker implements Callable<List<ITe
               TestResult.newEndTimeAwareTestResult(m_testMethod, m_testContext, null, start);
           r.setStatus(TestResult.SKIP);
           m_testResults.add(r);
-          m_invoker.runTestResultListener(r);
+          m_testInvoker.runTestResultListener(r);
           m_notifier.addSkippedTest(m_testMethod, r);
         }
       }
